use int instead of uint

+++ contracts/adapters/maker/ChaiTokenAdapter.sol
@@ -57,9 +57,11 @@ contract ChaiTokenAdapter is TokenAdapter, MKRAdapter {
         components[0] = Component({
             token: DAI,
             // solhint-disable-next-line not-rely-on-time
-            rate: mkrRmul(
-                mkrRmul(mkrRpow(pot.dsr(), block.timestamp - pot.rho(), ONE), pot.chi()),
-                1e18
+            rate: int256(
+                mkrRmul(
+                    mkrRmul(mkrRpow(pot.dsr(), block.timestamp - pot.rho(), ONE), pot.chi()),
+                    1e18
+                )
             )
         });
